add custom JWT token generation service to verify current user session

+++ pfm_expenses_server/src/main/java/iss/ibf/pfm_expenses_server/controller/AccountController.java
@@ -11,19 +11,23 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartFile;
 
+import iss.ibf.pfm_expenses_server.authentication.JwtTokenUtil;
 import iss.ibf.pfm_expenses_server.exception.NoEmailFoundException;
 import iss.ibf.pfm_expenses_server.exception.NoUserDetailsFoundException;
+import iss.ibf.pfm_expenses_server.exception.UnAuthorizedAccessException;
 import iss.ibf.pfm_expenses_server.exception.UsernameException;
 import iss.ibf.pfm_expenses_server.model.User;
 import iss.ibf.pfm_expenses_server.service.AccountService;
@@ -37,6 +41,7 @@ import jakarta.servlet.http.HttpSession;
 
 @Controller
 @RequestMapping(path={"/api"})
+@CrossOrigin("*")
 public class AccountController {
 
     @Autowired
@@ -50,6 +55,9 @@ public class AccountController {
 
     @Autowired
     private CurrencyConverterService currCnvSvc;
+
+    @Autowired
+    private JwtTokenUtil jwtTokenUtil;
 
     private Logger logger = Logger.getLogger(AccountController.class.getName());
     
@@ -104,11 +112,12 @@ public class AccountController {
     @PostMapping(path={"/account/login"}, consumes = MediaType.APPLICATION_JSON_VALUE)
     @ResponseBody
     public ResponseEntity<String> authenticateUser(@RequestBody String userLoginDetails, HttpSession session) {
+        
         JsonObject json = this.accSvc.convertStringToJsonObject(userLoginDetails);
 
         try {
             String username = json.getString("username");
-            session.setAttribute("username", username); // save current username to session
+            session.setAttribute("sessionUsername", username); // save current username to session
 
             // check if username and account are still valid
             if (this.accSvc.checkUsername(username)) {
@@ -129,11 +138,17 @@ public class AccountController {
                     try {
                         Boolean accCompleted = this.accSvc.checkAccountCompletion(username);
                         String email = this.accSvc.getUserEmail(username);
+
+                        // authenticate the user login by creating a security context
+                        String token = this.jwtTokenUtil.generateJwtToken(username);
+
                         JsonObject payload = Json.createObjectBuilder()
                                                     .add("accCompleted", accCompleted)
                                                     .add("accountId", accountId)
                                                     .add("email", email)
+                                                    .add("token", token)
                                                     .build();
+                                                    
                         return ResponseEntity.status(HttpStatus.OK).body(payload.toString());
 
                     } catch (NoUserDetailsFoundException ex) {
@@ -154,7 +169,7 @@ public class AccountController {
                         return ResponseEntity.status(HttpStatus.OK).body(payload.toString());
 
                     } catch (Exception ex) {
-
+                        System.out.println(">>> exception: " + ex.getMessage());
                         JsonObject error = Json.createObjectBuilder().add("error", ex.getMessage()).build();
 
                         return ResponseEntity.status(HttpStatus.FORBIDDEN).body(error.toString());
@@ -200,7 +215,12 @@ public class AccountController {
     // controller to complete user account
     @PostMapping(path={"/profile"}, consumes=MediaType.APPLICATION_JSON_VALUE)
     @ResponseBody
-    public ResponseEntity<String> completeUserInfo(@RequestBody String userInfoForm) {
+    public ResponseEntity<String> completeUserInfo(@RequestHeader("Authorization") String authorizationHeader, @RequestBody String userInfoForm, HttpSession session) {
+
+        if (!(this.jwtTokenUtil.getUserNameFromJwtToken(authorizationHeader.substring(7)).equals((String) session.getAttribute("sessionUsername")))) {
+
+            throw new UnAuthorizedAccessException("Access denied");
+        }
 
         JsonObject jsonUserInfo = this.accSvc.convertStringToJsonObject(userInfoForm);
 
@@ -215,18 +235,28 @@ public class AccountController {
                 return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(error.toString());
             }
 
+        } catch (UnAuthorizedAccessException ex) {
+
+            JsonObject error = Json.createObjectBuilder().add("error", ex.getMessage()).build();
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(error.toString());
+
         } catch (Exception ex) {
 
             JsonObject error = Json.createObjectBuilder().add("error", ex.getMessage()).build();
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(error.toString());
-        }
+        } 
     }
 
     // controller to retrieve user profile
     @GetMapping(path={"/profile"})
     @ResponseBody
-    public ResponseEntity<Object> retrieveProfile(@RequestParam String username, HttpSession session) {
+    public ResponseEntity<Object> retrieveProfile(@RequestHeader("Authorization") String authorizationHeader, @RequestParam String username, HttpSession session) {
         
+        // check if user session is valid
+        if (!(this.jwtTokenUtil.getUserNameFromJwtToken(authorizationHeader.substring(7)).equals((String) session.getAttribute("sessionUsername")))) {
+            throw new UnAuthorizedAccessException("Access denied");
+        }
+
         String accountId = (String) session.getAttribute("sessionAccountId");
         String userId = (String) session.getAttribute("sessionUserId");
 
@@ -234,8 +264,10 @@ public class AccountController {
             JsonObject payload = this.accSvc.getUserProfile(username, accountId, userId);
             return ResponseEntity.status(HttpStatus.OK).body(payload.toString());
 
+        } catch (UnAuthorizedAccessException ex) {
+            JsonObject error = Json.createObjectBuilder().add("error", ex.getMessage()).build();
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(error.toString());
         } catch (Exception ex) {
-
             JsonObject error = Json.createObjectBuilder().add("error", ex.getMessage()).build();
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(error.toString());
         }
@@ -244,7 +276,12 @@ public class AccountController {
     // controller to upload user profile picture
     @PostMapping(path={"/profile/image-upload"}, consumes=MediaType.MULTIPART_FORM_DATA_VALUE)
     @ResponseBody
-    public ResponseEntity<String> uploadProfilePicture(@RequestPart MultipartFile profilePic, @RequestPart String username, @RequestPart String accountId, HttpSession session) throws IOException {
+    public ResponseEntity<String> uploadProfilePicture(@RequestHeader("Authorization") String authorizationHeader, @RequestPart MultipartFile profilePic, @RequestPart String username, @RequestPart String accountId, HttpSession session) throws IOException {
+
+        // check if user session is valid
+        if (!(this.jwtTokenUtil.getUserNameFromJwtToken(authorizationHeader.substring(7)).equals((String) session.getAttribute("sessionUsername")))) {
+            throw new UnAuthorizedAccessException("Access denied");
+        }
 
         String userId = (String) session.getAttribute("sessionUserId");
 
@@ -253,7 +290,11 @@ public class AccountController {
             JsonObject payload = Json.createObjectBuilder().add("endpoint", endpoint).build();
             return ResponseEntity.status(HttpStatus.OK).body(payload.toString());
 
-        } catch (Exception ex) {
+        } catch (UnAuthorizedAccessException ex) {
+            JsonObject error = Json.createObjectBuilder().add("error", ex.getMessage()).build();
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(error.toString());
+        } 
+        catch (Exception ex) {
             JsonObject error = Json.createObjectBuilder().add("error", ex.getMessage()).build();
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(error.toString());
 
@@ -285,7 +326,12 @@ public class AccountController {
     // controller to insert new transaction
     @PostMapping(path={"/transaction/insert"})
     @ResponseBody
-    public ResponseEntity<String> insertTransaction(@RequestBody String form, HttpSession session) {
+    public ResponseEntity<String> insertTransaction(@RequestHeader("Authorization") String authorizationHeader, @RequestBody String form, HttpSession session) {
+
+        // check if user session is valid
+        if (!(this.jwtTokenUtil.getUserNameFromJwtToken(authorizationHeader.substring(7)).equals((String) session.getAttribute("sessionUsername")))) {
+            throw new UnAuthorizedAccessException("Access denied");
+        }
 
         String userId = (String) session.getAttribute("sessionUserId");
         JsonObject req = Json.createReader(new StringReader(form)).readObject();
@@ -300,6 +346,9 @@ public class AccountController {
                 JsonObject message = Json.createObjectBuilder().add("error", "Record insertion failed").build();
                 return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(message.toString());
             }
+        } catch (UnAuthorizedAccessException ex) {
+            JsonObject error = Json.createObjectBuilder().add("error", ex.getMessage()).build();
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(error.toString()); 
         } catch (Exception ex) {
             System.out.println(">>> exception: " + ex.getMessage());
             JsonObject error  = Json.createObjectBuilder().add("error", ex.getMessage()).build();
@@ -311,15 +360,22 @@ public class AccountController {
     // controller to retrieve transaction record by month
     @GetMapping(path={"/transaction/retrieve"})
     @ResponseBody
-    public ResponseEntity<String> retrieveTransactionRecordByMonth(@RequestParam String startDate, HttpSession session) {
+    public ResponseEntity<String> retrieveTransactionRecordByMonth(@RequestHeader("Authorization") String authorizationHeader, @RequestParam String startDate, HttpSession session) {
 
-        System.out.println(">>> retrieving... ");
+        // check if user session is valid
+        if (!(this.jwtTokenUtil.getUserNameFromJwtToken(authorizationHeader.substring(7)).equals((String) session.getAttribute("sessionUsername")))) {
+            throw new UnAuthorizedAccessException("Access denied");
+        }
+
         String userId = (String) session.getAttribute("sessionUserId");
 
         try {
             JsonObject payload = this.transSvc.retrieveTransaction(startDate, userId);
             return ResponseEntity.status(HttpStatus.OK).body(payload.toString());
 
+        } catch (UnAuthorizedAccessException ex) {
+            JsonObject error = Json.createObjectBuilder().add("error", ex.getMessage()).build();
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(error.toString()); 
         } catch (Exception ex) {
             System.out.println(">>> exception: " + ex.getMessage());
             JsonObject error = Json.createObjectBuilder().add("error", ex.getMessage()).build();
@@ -331,11 +387,15 @@ public class AccountController {
     // controller to update transaction record by id
     @PutMapping(path={"/transaction/update"})
     @ResponseBody
-    public ResponseEntity<String> updateTransactionRecordById(@RequestBody String editForm, HttpSession session) {
+    public ResponseEntity<String> updateTransactionRecordById(@RequestHeader("Authorization") String authorizationHeader, @RequestBody String editForm, HttpSession session) {
+
+        // check if user session is valid
+        if (!(this.jwtTokenUtil.getUserNameFromJwtToken(authorizationHeader.substring(7)).equals((String) session.getAttribute("sessionUsername")))) {
+            throw new UnAuthorizedAccessException("Access denied");
+        }
 
         String userId = (String) session.getAttribute("sessionUserId");
         JsonObject record = Json.createReader(new StringReader(editForm)).readObject();
-        System.out.println(">>> request: " + record.toString());
 
         try {
             if (this.transSvc.updateTransaction(record, userId)) {
@@ -345,6 +405,9 @@ public class AccountController {
                 JsonObject fail = Json.createObjectBuilder().add("fail", "Update failed").build();
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(fail.toString());
             }
+        } catch (UnAuthorizedAccessException ex) {
+            JsonObject error = Json.createObjectBuilder().add("error", ex.getMessage()).build();
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(error.toString());  
         } catch (Exception ex) {
             System.out.println(">>> exception while updating: " + ex.getMessage());
             JsonObject error = Json.createObjectBuilder().add("error", ex.getMessage()).build();
@@ -355,7 +418,12 @@ public class AccountController {
     // controller to delete transaction record by id
     @DeleteMapping(path={"/transaction/delete"})
     @ResponseBody
-    public ResponseEntity<String> deleteTransactionRecordById(@RequestParam Integer transactionId, HttpSession session) {
+    public ResponseEntity<String> deleteTransactionRecordById(@RequestHeader("Authorization") String authorizationHeader, @RequestParam Integer transactionId, HttpSession session) {
+
+        // check if user session is valid
+        if (!(this.jwtTokenUtil.getUserNameFromJwtToken(authorizationHeader.substring(7)).equals((String) session.getAttribute("sessionUsername")))) {
+            throw new UnAuthorizedAccessException("Access denied");
+        }
 
         try {
             if (this.transSvc.deleteTransaction(transactionId)) {
@@ -365,6 +433,9 @@ public class AccountController {
                 JsonObject fail = Json.createObjectBuilder().add("fail", "Delete failed").build();
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(fail.toString());
             }
+        } catch (UnAuthorizedAccessException ex) {
+            JsonObject error = Json.createObjectBuilder().add("error", ex.getMessage()).build();
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(error.toString());   
         } catch (Exception ex) {
             System.out.println(">>> exception during deleting: " + ex.getMessage());
             JsonObject error = Json.createObjectBuilder().add("error", ex.getMessage()).build();
